+++ frontend/components/pages/Home.js
@@ -3,16 +3,15 @@
  * See the license file in the root folder for details.
  */
 import React, { useState } from 'react';
-import { useHistory } from 'react-router-dom';
 import cx from 'classnames';
 import { useQueryParam, StringParam } from 'use-query-params';
-import SearchBar from '../ResultsPage/SearchBar';
 import logo from '../images/logo.svg';
 import boston from '../images/boston.svg';
 import macros from '../macros';
 import SplashPage from '../SplashPage/SplashPage';
 import Footer from '../Footer';
-import TermDropdown, { termDropDownOptions } from '../ResultsPage/TermDropdown';
+import { termDropDownOptions } from '../ResultsPage/TermDropdown';
+import HomeSearch from '../HomePage/HomeSearch'
 
 
 const ATTENTION_SECTION = {
@@ -28,7 +27,6 @@ const LATEST_TERM = '202110';
 const AVAILABLE_TERMS = termDropDownOptions.map((t) => { return t.value; });
 
 export default function Home() {
-  const history = useHistory();
   const [termId = LATEST_TERM, setTermId] = useQueryParam('termId', StringParam); // Default to LATEST if term not in params
   // Redirect to latest if we're at an old term
   if (!AVAILABLE_TERMS.includes(termId)) {
@@ -91,21 +89,7 @@ export default function Home() {
         >
           <div className='centerTextContainer'>
             <img src={ logo } className='logo' alt='logo' />
-            <div
-              className='searchWrapper'
-              onFocus={ () => { setSearchFocused(true); } }
-              onBlur={ () => { setSearchFocused(false); } }
-            >
-              <SearchBar
-                onSearch={ (q) => { history.push(`/${termId}/${q}`); } }
-                query=''
-              />
-            </div>
-            <TermDropdown
-              termId={ termId }
-              onChange={ setTermId }
-            />
-
+            <HomeSearch setSearchFocused={ setSearchFocused } setTermId={ setTermId } termId={ termId } />
             <div className={ cx({
               hitEnterToSearch: true,
               hitEnterToSearchVisible: searchFocused,
