+++ frontend/css/_TermDropDown.scss
@@ -9,7 +9,7 @@
   pointer-events: initial;
 
   & * {
-    font-size: 18px;
+    font-size: 17px;
     font-weight: initial !important;
     border: none !important;
     color: #636363 !important;
