+++ frontend/css/base.scss
@@ -35,6 +35,7 @@
 @import 'results/MobileSearchResult';
 @import 'results/MobileSectionPanel';
 @import 'results/NotifCheckBox';
+@import 'home/HomeSearch';
 
 html,
 body,
